fix getting userid

+++ routes/subscription.ts
@@ -5,6 +5,8 @@ import {
   removeNotification,
   scheduleNotification,
 } from '../utils/subscription';
+import { getUser } from '../utils/getUser';
+import { saveLog } from '../logger';
 
 const router = express.Router();
 
@@ -19,9 +21,20 @@ router.get('/vapidPublicKey', (req, res) => {
 });
 
 router.post('/subscribe', async (req, res) => {
+  const user = await getUser(req?.headers?.authorization);
+
+  if (user === 401 || !user) {
+    saveLog('error', 'GET', 'subscription/subscribe', 'no logged user', {
+      user,
+    });
+    res.status(404).json({ message: 'no logged user' });
+    return;
+  }
+
+  const userId = user?._id;
   const subscription = new SubscriptionModel({
     ...req.body.subscription,
-    userId: req.body.userId,
+    userId,
   });
 
   try {
@@ -34,8 +47,18 @@ router.post('/subscribe', async (req, res) => {
   }
 });
 
-router.post('/unsubscribe', async (req, res) => {
-  const { userId } = req.body;
+router.delete('/unsubscribe', async (req, res) => {
+  const user = await getUser(req?.headers?.authorization);
+
+  if (user === 401 || !user) {
+    saveLog('error', 'GET', 'subscription/unsubscribe', 'no logged user', {
+      user,
+    });
+    res.status(404).json({ message: 'no logged user' });
+    return;
+  }
+
+  const userId = user?._id.toString();
 
   try {
     await SubscriptionModel.deleteMany({ userId });
@@ -47,29 +70,29 @@ router.post('/unsubscribe', async (req, res) => {
   }
 });
 
-router.post('/sendNotification', async (req, res) => {
-  const notificationPayload = JSON.stringify({
-    title: req.body.title,
-    body: req.body.body,
-  });
+// router.post('/sendNotification', async (req, res) => {
+//   const notificationPayload = JSON.stringify({
+//     title: req.body.title,
+//     body: req.body.body,
+//   });
 
-  try {
-    const subscriptions = await SubscriptionModel.find({
-      userId: req.body.userId,
-    });
+//   try {
+//     const subscriptions = await SubscriptionModel.find({
+//       userId: req.body.userId,
+//     });
 
-    const promises = subscriptions.map((sub) => {
-      console.log(2, { sub });
-      webpush.sendNotification(sub.toJSON(), notificationPayload);
-    });
-    console.log(1, { promises });
+//     const promises = subscriptions.map((sub) => {
+//       console.log(2, { sub });
+//       webpush.sendNotification(sub.toJSON(), notificationPayload);
+//     });
+//     console.log(1, { promises });
 
-    await Promise.all(promises);
-    res.status(200).json({ message: 'Notifications sent' });
-  } catch (error) {
-    console.error('Error sending notification, reason: ', error);
-    res.sendStatus(500);
-  }
-});
+//     await Promise.all(promises);
+//     res.status(200).json({ message: 'Notifications sent' });
+//   } catch (error) {
+//     console.error('Error sending notification, reason: ', error);
+//     res.sendStatus(500);
+//   }
+// });
 
 export default router;
